python-climate-ews/requirements.txt
# NOTE: This project targets Python 3.11.x (see runtime.txt). Using newer Python
# (e.g. 3.12+ / 3.14) may fail because TensorFlow and compiled dependencies may
# not have wheels available and would require building from source.
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
requests==2.31.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
statsmodels==0.14.2
